+++ R/import.R
@@ -1,6 +1,6 @@
 #' @importFrom dplyr     %>% arrange bind_rows contains filter mutate pull
 #'                       rename select
 #' @importFrom readr     col_character col_date col_datetime col_double
-#'                       col_integer cols problems read_csv
+#'                       col_integer cols locale problems read_csv
 #' @importFrom rlang    .data
 NULL